+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add creation of new os group `shinesolutions`
 - Add rsyslog configuration file for file/dir permissions
 - Introducing new configuration parameter to install AWS Kinesis agent
-
+- Update group permissions for `/var/log` to groupowner `shinesolutions`
 
 ## 4.11.1 - 2019-09-12